requirements.txt
amqp==2.5.2
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.2
cffi==1.14.0
coverage==5.0.4
Django==3.0.5
django-debug-toolbar==2.2
djangorestframework==3.11.0
gevent==1.4.0
greenlet==0.4.15
kombu==4.6.8
Pillow==7.1.1
pycparser==2.20
pytz==2019.3
redis==3.4.1
sqlparse==0.3.1
vine==1.3.0
